+++ config.go
@@ -31,6 +31,12 @@ func GenerateConfig(containers []containuum.Container, routeExtras string) strin
 		}
 		sb.WriteString("\n")
 
+		if route.Provider != "" {
+			sb.WriteString("    provider ")
+			sb.WriteString(route.Provider)
+			sb.WriteString("\n")
+		}
+
 		for _, upstream := range route.Upstreams {
 			sb.WriteString("    upstream ")
 			sb.WriteString(upstream.Name)